backend/projects/news_aggregator/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
pyyaml==6.0.1
pytest==8.3.2
pytest-asyncio==0.23.7
anyio==4.4.0
